+++ CHANGELOG.md
@@ -3,6 +3,7 @@ We use [semantic versioning][semver]
 # Next Release
 - [fix] `git.properties` commit was not used for upload
 - [feature] Upload ignored `origin/` prefix in `git.properties`'s branch name
+- [fix] Executable spring boot jar produces no coverage
 
 # 15.3.0
 - [feature] Added `coverage-via-http` option for testwise mode